+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - enabled dependabot
+- zconfigcheck integration is now validated by GitHub Actions
 
 ## 0.2.0 - 2024-09-05
 ### Changed